requirements.txt
streamlit
pandas
pillow
openpyxl